Refactor - remove new cicrcular dependecy
